% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{check_api_status}
\alias{check_api_status}
\title{Check the MLFlow API status}
\usage{
check_api_status(client = NULL)
}
\arguments{
\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\value{
"ok" if the API is healthy, otherwise returns an HTTP error.
}
\description{
Check the MLFlow API status
}
