% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{create_nodelete_tag}
\alias{create_nodelete_tag}
\title{Create a NODELETE tag for an experiment}
\usage{
create_nodelete_tag(experiment_id)
}
\arguments{
\item{experiment_id}{The experiment ID for which to set the NODELETE tag}
}
\value{
No return value. Called for side effects.
}
\description{
We create \code{NODELETE} tags for experiments that we, uh, don't want to delete.
\code{NODELETE} tags will only be set in either non-interactive sessions or by explicit request.
}
