% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{create_registered_model}
\alias{create_registered_model}
\title{Create registered model}
\usage{
create_registered_model(
  name,
  tags = list(),
  description = "",
  client = mlflow_client()
)
}
\arguments{
\item{name}{The name of the model to create.}

\item{tags}{Additional metadata for the registered model (Optional).}

\item{description}{Description for the registered model (Optional).}

\item{client}{An MLFlow client. Will be auto-generated if omitted.}
}
\value{
A list of metadata on the registered model, including the \code{name} and the \code{creation_timestamp} and \code{last_updated_timestamp}
}
\description{
Creates a new registered model in the model registry
}
