% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{delete_experiment}
\alias{delete_experiment}
\title{Delete Experiment}
\usage{
delete_experiment(experiment_id, client = mlflow_client())
}
\arguments{
\item{experiment_id}{ID of the associated experiment. This field is required.}

\item{client}{An MLFlow client. If not provided, the client is sourced from the \code{MLFLOW_TRACKING_URI} environment variable.}
}
\value{
No return value. Called for side effects.
}
\description{
Marks an experiment and associated runs, params, metrics, etc. for deletion. If the
experiment uses FileStore, artifacts associated with experiment are also deleted.
}
