% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{delete_run}
\alias{delete_run}
\title{Delete a Run}
\usage{
delete_run(run_id = get_active_run_id(), client = mlflow_client())
}
\arguments{
\item{run_id}{A run uuid. Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\value{
No return value. Called for side effects.
}
\description{
Deletes the run with the specified ID.
}
