% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{delete_tag}
\alias{delete_tag}
\title{Delete Tag}
\usage{
delete_tag(key, run_id = get_active_run_id(), client = mlflow_client())
}
\arguments{
\item{key}{Name of the tag. Maximum size is 255 bytes. This field is required.}

\item{run_id}{A run uuid. Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\description{
Deletes a tag on a run. This is irreversible. Tags are run metadata that can be updated during a run and
after a run completes.
}
