% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{download_artifact}
\alias{download_artifact}
\title{Download Artifact}
\usage{
download_artifact(path, run_id = get_active_run_id(), ...)
}
\arguments{
\item{path}{Relative source path to the desired artifact.}

\item{run_id}{A run uuid. Automatically inferred if a run is currently active.}

\item{...}{Additional arguments to pass to \code{aws.s3::save_object}}
}
\description{
Download an artifact file or directory from a run to a local directory if applicable,
and return a local path for it.
}
