% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{end_run}
\alias{end_run}
\title{End a Run}
\usage{
end_run(
  status = c("FINISHED", "FAILED", "KILLED"),
  run_id = get_active_run_id(),
  client = mlflow_client()
)
}
\arguments{
\item{status}{Updated status of the run. Defaults to \code{FINISHED}. Can also be set to
"FAILED" or "KILLED".}

\item{run_id}{A run uuid. Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\value{
Metadata on the newly-ended run.
}
\description{
Terminates a run. Attempts to end the current active run if \code{run_id} is not specified.
}
