% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{exists_active_run}
\alias{exists_active_run}
\title{Check if an active run exists}
\usage{
exists_active_run()
}
\value{
\code{TRUE} if an active run exists, \code{FALSE} otherwise
}
\description{
Check if an active run exists
}
