% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{gather_model_metadata}
\alias{gather_model_metadata}
\title{Gather metadata for a model version}
\usage{
gather_model_metadata(registered_model_name, version, client = mlflow_client())
}
\arguments{
\item{registered_model_name}{The name of the registered model}

\item{version}{The version of the model for which to gather metadata}

\item{client}{An MLFlow client. Autogenerated if not provided}
}
\value{
A list containing the \code{registered_model_name}, the \code{experiment_id}, the \code{experiment_name}, the \code{run_id}, and the \code{model_version}
}
\description{
Model versions are tied to experiments, runs, and registered model names.
This helper returns metadata on the experiment and run that are associated with the registered model name and version provided.
}
