% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{get_active_experiment_id}
\alias{get_active_experiment_id}
\title{Get the ID of the active experiment}
\usage{
get_active_experiment_id()
}
\value{
The active experiment ID
}
\description{
Get the ID of the active experiment
}
