% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{get_artifact_path}
\alias{get_artifact_path}
\title{Get the artifact path for a run}
\usage{
get_artifact_path(run_id = get_active_run_id(), client = mlflow_client())
}
\arguments{
\item{run_id}{A run id. Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Auto-generated if not provided.}
}
\value{
A path to the run's artifacts in S3
}
\description{
Get the artifact path for a run
}
