% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{get_experiment}
\alias{get_experiment}
\title{Get Experiment}
\usage{
get_experiment(
  experiment_id = get_active_experiment_id(),
  experiment_name = NULL,
  client = mlflow_client()
)
}
\arguments{
\item{experiment_id}{ID of the experiment.}

\item{experiment_name}{The experiment name. Only one of \code{experiment_name} or \code{experiment_id} should be specified.}

\item{client}{An MLFlow client. If not provided, the client is sourced from the \code{MLFLOW_TRACKING_URI} environment variable.}
}
\value{
A tibble with metadata on the experiment requested.
}
\description{
Gets metadata for an experiment and a list of runs for the experiment. Attempts to obtain the
active experiment if both \code{experiment_id} and \code{experiment_name} are unspecified.
}
