% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{get_metric}
\alias{get_metric}
\title{Get the most recent value of a metric for a run}
\usage{
get_metric(metric_key, run_id = get_active_run_id(), client = mlflow_client())
}
\arguments{
\item{metric_key}{The metric key to get}

\item{run_id}{The run ID to get the metric for. Defaults to the active run}

\item{client}{An MLFlow client. Autogenerated if missing.}
}
\value{
The value of the metric
}
\description{
Get the most recent value of a metric for a run
}
