% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{get_metric_history}
\alias{get_metric_history}
\title{Get Metric History}
\usage{
get_metric_history(
  metric_key,
  run_id = get_active_run_id(),
  client = mlflow_client()
)
}
\arguments{
\item{metric_key}{Name of the metric.}

\item{run_id}{A run uuid. Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\description{
Get a list of all values for the specified metric for a given run.
}
