% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{get_param}
\alias{get_param}
\title{Get a parameter for a run}
\usage{
get_param(param, run_id = get_active_run_id(), client = mlflow_client())
}
\arguments{
\item{param}{The parameter to get}

\item{run_id}{The run ID to get the metric for. Defaults to the active run}

\item{client}{An MLFlow client. Autogenerated if missing.}
}
\value{
The value of the metric
}
\description{
Get a parameter for a run
}
