% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{get_registered_model}
\alias{get_registered_model}
\title{Get a registered model}
\usage{
get_registered_model(name, client = NULL)
}
\arguments{
\item{name}{The name of the model to retrieve.}
}
\description{
Retrieves a registered model from the Model Registry.
}
