% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{get_registered_model_run}
\alias{get_registered_model_run}
\title{Get a registered model run}
\usage{
get_registered_model_run(
  model_name,
  client = mlflow_client(),
  stage = "Production"
)
}
\arguments{
\item{model_name}{A model name.}

\item{client}{An MLFlow client. Will be auto-generated if omitted.}

\item{stage}{A model stage. Set to \code{NULL} or \code{NA} to return all results.}
}
\description{
Get a registered model run
}
