% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{get_registered_model_run_id}
\alias{get_registered_model_run_id}
\title{Get a registered model run id}
\usage{
get_registered_model_run_id(
  experiment_name,
  client = mlflow_client(),
  stage = "Production"
)
}
\arguments{
\item{experiment_name}{An experiment name.}

\item{client}{An MLFlow client. Will be auto-generated if omitted.}

\item{stage}{A model stage. Set to \code{NULL} to return all results.}
}
\description{
Get a registered model run id
}
