% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{get_run}
\alias{get_run}
\title{Get Run}
\usage{
get_run(run_id = get_active_run_id(), client = mlflow_client())
}
\arguments{
\item{run_id}{A run uuid. Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\value{
Metadata on the run, including the ID, the \code{experiment_id}, the \code{status}, etc.
}
\description{
Gets metadata, params, tags, and metrics for a run. Returns a single value for each metric
key: the most recently logged metric value at the largest step.
}
