% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{get_tracking_uri}
\alias{get_tracking_uri}
\title{Get Remote Tracking URI}
\usage{
get_tracking_uri()
}
\value{
The tracking URI
}
\description{
Gets the remote tracking URI. If no global is specified, defaults to the \code{MLFLOW_TRACKING_URI} environment variable.
}
