% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightMLFlow-package.R
\docType{package}
\name{lightMLFlow-package}
\alias{lightMLFlow}
\alias{lightMLFlow-package}
\title{lightMLFlow: A Lightweight R Wrapper for the MLFlow REST API}
\description{
MLFlow is an open-source tool for tracking, versioning, and deploying machine learning models. This package provides a lightweight R wrapper for the MLFlow REST API. It makes some assumptions (notably, that you already have an MLFlow instance running, you want to use S3 for artifact storage, and you intend to deploy models in an (e.g.) `plumber` API), but, in return, you get some benefits: A simpler API, not needing 'reticulate', relying on 'aws.s3' where possible, etc.
}
\author{
\strong{Maintainer}: Matt Kaye \email{mrkaye97@gmail.com}

Authors:
\itemize{
  \item Tony ElHabr \email{anthonyelhabr@gmail.com}
}

}
\keyword{internal}
