% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{list_experiments}
\alias{list_experiments}
\title{List Experiments}
\usage{
list_experiments(
  view_type = c("ACTIVE_ONLY", "DELETED_ONLY", "ALL"),
  client = mlflow_client()
)
}
\arguments{
\item{view_type}{Qualifier for type of experiments to be returned. Defaults to \code{ACTIVE_ONLY}.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\description{
Gets a list of all experiments.
}
