% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{list_registered_models}
\alias{list_registered_models}
\title{List registered models}
\usage{
list_registered_models(max_results = 100, page_token = NULL, client)
}
\arguments{
\item{max_results}{Maximum number of registered models to retrieve.}

\item{page_token}{Pagination token to go to the next page based on a
previous query.}
}
\description{
Retrieves a list of registered models.
}
