% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{load_artifact}
\alias{load_artifact}
\title{Load an artifact into an R object}
\usage{
load_artifact(
  artifact_name,
  FUN = readRDS,
  run_id = get_active_run_id(),
  client = mlflow_client(),
  pause_base = 0.5,
  max_times = 5,
  pause_cap = 60,
  ...
)
}
\arguments{
\item{artifact_name}{The name of the artifact to load}

\item{FUN}{a function to use to load the artifact}

\item{run_id}{A run ID to find the URI for}

\item{client}{An MLFlow client}

\item{pause_base, max_times, pause_cap}{See \link[purrr]{insistently}}

\item{\dots}{Additional arguments to pass on to \code{s3read_using}}
}
\value{
An R object. The result of \code{s3read_using}
}
\description{
Load an artifact into an R object
}
