% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{load_model}
\alias{load_model}
\title{Load MLflow Model}
\usage{
load_model(model_uri, flavor = NULL, client = mlflow_client())
}
\arguments{
\item{model_uri}{The URI to the model to load.}

\item{flavor}{Optional flavor specification (string). Can be used to load a particular flavor in
case there are multiple flavors available.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\description{
Loads an MLflow model. MLflow models can have multiple model flavors. Not all flavors / models
can be loaded in R. This method by default searches for a flavor supported by R/MLflow.
}
