% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{log_artifact}
\alias{log_artifact}
\title{Log Artifact}
\usage{
log_artifact(file, run_id = NULL, client = NULL, ...)
}
\arguments{
\item{file}{The file to log as an artifact.}

\item{run_id}{A run uuid. Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}

\item{...}{Additional arguments to pass to \code{aws.s3::put_object}}
}
\description{
Logs a specific file or directory as an artifact for a run.
}
\details{
When logging to Amazon S3, ensure that you have the s3:PutObject, s3:GetObject,
s3:ListBucket, and s3:GetBucketLocation permissions on your bucket.

Additionally, at least the \code{AWS_ACCESS_KEY_ID} and \code{AWS_SECRET_ACCESS_KEY}
environment variables must be set to the corresponding key and secrets provided
by Amazon IAM.
}
