% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{log_batch}
\alias{log_batch}
\title{Log Batch}
\usage{
log_batch(
  metrics = NULL,
  params = NULL,
  tags = list(),
  run_id = NULL,
  client = NULL
)
}
\arguments{
\item{metrics}{A dataframe of metrics to log, containing the following columns: "key", "value",
"step", "timestamp". This dataframe cannot contain any missing ('NA') entries.}

\item{params}{A dataframe of params to log, containing the following columns: "key", "value".
This dataframe cannot contain any missing ('NA') entries.}

\item{tags}{A dataframe of tags to log, containing the following columns: "key", "value".
This dataframe cannot contain any missing ('NA') entries.}

\item{run_id}{A run uuid. Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\description{
Log a batch of metrics, params, and/or tags for a run. The server will respond with an error (non-200 status code)
if any data failed to be persisted. In case of error (due to internal server error or an invalid request), partial
data may be written.
}
