% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{log_metrics}
\alias{log_metrics}
\title{Log Metrics}
\usage{
log_metrics(..., run_id = get_active_run_id(), client = mlflow_client())
}
\arguments{
\item{...}{variable names from which a data.frame with \code{key} and \code{value} columns will be created.}

\item{run_id}{A run uuid. Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\value{
No return value. Called for side effects.
}
\description{
Logs a metric for a run. Metrics key-value pair that records a single float measure.
During a single execution of a run, a particular metric can be logged several times.
The MLflow Backend keeps track of historical metric values along two axes: timestamp and step.
}
