% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{log_model}
\alias{log_model}
\title{Log Model}
\usage{
log_model(model, path, run_id = get_active_run_id(), ...)
}
\arguments{
\item{model}{The model that will perform a prediction.}

\item{path}{Destination path where this MLflow compatible model
will be saved.}

\item{run_id}{The run id to log the model under. Defaults to the active run.}

\item{...}{Optional additional arguments passed to \code{save_model()} when persisting the
model. For example, \verb{conda_env = /path/to/conda.yaml} may be passed to specify a conda
dependencies file for flavors (e.g. keras) that support conda environments.}
}
\description{
Logs a model for this run. Similar to \code{save_model()}
but stores model as an artifact within the active run.
}
