% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{log_param}
\alias{log_param}
\title{Log Parameter}
\usage{
log_param(key, value, run_id, client)
}
\arguments{
\item{key}{Name of the parameter.}

\item{value}{String value of the parameter.}

\item{run_id}{A run uuid. Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\description{
Logs a parameter for a run. Examples are params and hyperparams
used for ML training, or constant dates and values used in an ETL pipeline.
A param is a STRING key-value pair. For a run, a single parameter is allowed
to be logged only once.
}
