% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{log_params}
\alias{log_params}
\title{Log Parameters}
\usage{
log_params(..., run_id, client)
}
\description{
Logs parameters for a run. Examples are params and hyperparams
used for ML training, or constant dates and values used in an ETL pipeline.
A param is a STRING key-value pair. For a run, a single parameter is allowed
to be logged only once.
}
