% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mlflow_id}
\alias{mlflow_id}
\alias{mlflow_id.mlflow_run}
\alias{mlflow_id.mlflow_experiment}
\title{Get Run or Experiment ID}
\usage{
mlflow_id(object)

\method{mlflow_id}{mlflow_run}(object)

\method{mlflow_id}{mlflow_experiment}(object)
}
\arguments{
\item{object}{An \code{mlflow_run} or \code{mlflow_experiment} object.}
}
\description{
Extracts the ID of the run or experiment.
}
