% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{record_logged_model}
\alias{record_logged_model}
\title{Record logged model metadata with the tracking server.}
\usage{
record_logged_model(model_spec, run_id = NULL, client = NULL)
}
\arguments{
\item{model_spec}{A model specification.}

\item{run_id}{A run uuid. Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\description{
Record logged model metadata with the tracking server.
}
