% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{register_external_observer}
\alias{register_external_observer}
\title{Register an external MLflow observer}
\usage{
register_external_observer(observer)
}
\arguments{
\item{observer}{The observer object (see example)}
}
\description{
Registers an external MLflow observer that will receive a
\code{register_tracking_event(event_name, data)} callback on any model tracking
event such as "create_run", "delete_run", or "log_metric".
Each observer should have a \code{register_tracking_event(event_name, data)}
callback accepting a character vector \code{event_name} specifying the name of
the tracking event, and \code{data} containing a list of attributes of the event.
The callback should be non-blocking, and ideally should complete
instantaneously. Any exception thrown from the callback will be ignored.
}
