% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{rename_experiment}
\alias{rename_experiment}
\title{Rename Experiment}
\usage{
rename_experiment(
  new_name,
  experiment_id = get_active_experiment_id(),
  client = mlflow_client()
)
}
\arguments{
\item{new_name}{The experiment's name will be changed to this. The new name must be unique.}

\item{experiment_id}{ID of the associated experiment. This field is required.}

\item{client}{An MLFlow client. If not provided, the client is sourced from the \code{MLFLOW_TRACKING_URI} environment variable.}
}
\value{
No return value. Called for side effects.
}
\description{
Renames an experiment.
}
