% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{restore_experiment}
\alias{restore_experiment}
\title{Restore Experiment}
\usage{
restore_experiment(experiment_id, client = mlflow_client())
}
\arguments{
\item{experiment_id}{ID of the associated experiment. This field is required.}

\item{client}{An MLFlow client. If not provided, the client is sourced from the \code{MLFLOW_TRACKING_URI} environment variable.}
}
\description{
Restores an experiment marked for deletion. This also restores associated metadata,
runs, metrics, and params. If experiment uses FileStore, underlying artifacts
associated with experiment are also restored.
}
\details{
Throws \code{RESOURCE_DOES_NOT_EXIST} if the experiment was never created or was permanently deleted.
}
