% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{restore_run}
\alias{restore_run}
\title{Restore a Run}
\usage{
restore_run(run_id, client)
}
\arguments{
\item{run_id}{A run id}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\description{
Restores the run with the specified ID.
}
