% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{s3_select_from_artifact}
\alias{s3_select_from_artifact}
\title{Call the S3 SELECT API on a CSV artifact}
\usage{
s3_select_from_artifact(
  artifact_name,
  run_id = get_active_run_id(),
  client = mlflow_client(),
  pause_base = 0.5,
  max_times = 5,
  pause_cap = 60,
  Expression,
  ExpressionType = "SQL",
  InputSerialization = list(CSV = list(FileHeaderInfo = "NONE", RecordDelimiter =
    "\\n", FieldDelimiter = ","), CompressionType = "NONE"),
  OutputSerialization = list(CSV = list(RecordDelimiter = "\\n", FieldDelimiter = ",",
    QuoteCharacter = "\\"", QuoteFields = "ASNEEDED")),
  ...
)
}
\arguments{
\item{artifact_name}{The name of the artifact to \code{SELECT} from}

\item{run_id}{A run id Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}

\item{pause_base, max_times, pause_cap}{See \link[purrr]{insistently}}

\item{Expression, ExpressionType, InputSerialization, OutputSerialization}{See \link[paws.storage]{s3_select_object_content}}

\item{\dots}{Additional arguments to pass to \link[paws.storage]{s3_select_object_content}}
}
\value{
A data.frame of the query result
}
\description{
Call the S3 SELECT API on a CSV artifact
}
