% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-crate.R, R/model.R
\name{save_model.crate}
\alias{save_model.crate}
\alias{save_model}
\title{Save Model for MLflow}
\usage{
\method{save_model}{crate}(model, path, model_spec = list(), ...)

save_model(model, path, model_spec = list(), ...)
}
\arguments{
\item{model}{The model that will perform a prediction.}

\item{path}{Destination path where this MLflow compatible model
will be saved.}

\item{model_spec}{MLflow model config this model flavor is being added to.}

\item{...}{Additional arguments to pass to \code{log_artifact}}
}
\description{
Saves model in MLflow format that can later be used for prediction and serving. This method is
generic to allow package authors to save custom model types.
}
