% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{search_artifacts}
\alias{search_artifacts}
\alias{list_artifacts}
\title{List Artifacts}
\usage{
search_artifacts(
  path = NULL,
  run_id = get_active_run_id(),
  client = mlflow_client()
)

list_artifacts(
  path = NULL,
  run_id = get_active_run_id(),
  client = mlflow_client()
)
}
\arguments{
\item{path}{The run's relative artifact path to list from. If not specified, it is
set to the root artifact path}

\item{run_id}{A run id Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\value{
A \code{data.frame} of the artifacts at the path provided for the run provided.
}
\description{
Gets a list of artifacts.
}
