% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{search_experiments}
\alias{search_experiments}
\alias{list_experiments}
\title{List Experiments}
\usage{
search_experiments(
  view_type = c("ACTIVE_ONLY", "DELETED_ONLY", "ALL"),
  client = mlflow_client()
)

list_experiments(
  view_type = c("ACTIVE_ONLY", "DELETED_ONLY", "ALL"),
  client = mlflow_client()
)
}
\arguments{
\item{view_type}{Qualifier for type of experiments to be returned. Defaults to \code{ACTIVE_ONLY}.}

\item{client}{An MLFlow client. If not provided, the client is sourced from the \code{MLFLOW_TRACKING_URI} environment variable.}
}
\value{
A \code{data.frame} of experiments, with columns \code{experiment_id}, \code{name}, \code{artifact_location}, \code{lifecycle_stage}, and \code{tags}
}
\description{
Gets a list of all experiments.
}
