% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{search_runs}
\alias{search_runs}
\title{Search Runs}
\usage{
search_runs(
  experiment_ids,
  run_view_type = c("ACTIVE_ONLY", "DELETED_ONLY", "ALL"),
  order_by = list(),
  filter = "",
  client = mlflow_client()
)
}
\arguments{
\item{experiment_ids}{List of string experiment IDs (or a single string experiment ID) to search
over. Attempts to use active experiment if not specified.}

\item{run_view_type}{Run view type.}

\item{order_by}{List of properties to order by. Example: "metrics.acc DESC".}

\item{filter}{A filter expression over params, metrics, and tags, allowing returning a subset of runs.
The syntax is a subset of SQL which allows only ANDing together binary operations between a param/metric/tag and a constant.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\value{
A data.frame of runs matching the search criteria.
}
\description{
Search for runs that satisfy expressions. Search expressions can use Metric and Param keys.
}
