% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{set_active_experiment_id}
\alias{set_active_experiment_id}
\title{Set an experiment to \code{active}}
\usage{
set_active_experiment_id(experiment_id)
}
\arguments{
\item{experiment_id}{The ID of the experiment to activate}
}
\value{
No return value. Called for side effects.
}
\description{
Set an experiment to \code{active}
}
