% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{set_tag}
\alias{set_tag}
\title{Set Tag}
\usage{
set_tag(key, value, run_id = NULL, client = NULL)
}
\arguments{
\item{key}{Name of the tag. Maximum size is 255 bytes. This field is required.}

\item{value}{String value of the tag being logged. Maximum size is 500 bytes. This field is required.}

\item{run_id}{A run uuid. Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\description{
Sets a tag on a run. Tags are run metadata that can be updated during a run and
after a run completes.
}
