% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{start_run}
\alias{start_run}
\title{Start Run}
\usage{
start_run(
  run_id = Sys.getenv("MLFLOW_RUN_ID"),
  experiment_id = get_active_experiment_id(),
  client = mlflow_client(),
  nested = FALSE
)
}
\arguments{
\item{run_id}{If specified, get the run with the specified UUID and log metrics
and params under that run. The run's end time is unset and its status is set to
running, but the run's other attributes remain unchanged.}

\item{experiment_id}{Used only when \code{run_id} is unspecified. ID of the experiment under
which to create the current run. If unspecified, the run is created under
a new experiment with a randomly generated name.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}

\item{nested}{Controls whether the run to be started is nested in a parent run. \code{TRUE} creates a nest run.}
}
\value{
Metadata on the newly-started run.
}
\description{
Starts a new run. If \code{client} is not provided, this function infers contextual information such as
source name and version, and also registers the created run as the active run. If \code{client} is provided,
no inference is done.
}
