% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_for_missing_columns}
\alias{stop_for_missing_columns}
\title{Abort on column missingness}
\usage{
stop_for_missing_columns(columns, data)
}
\arguments{
\item{columns}{a character vector of expected columns}

\item{data}{a data frame containing the expected columns}
}
\value{
No return value. Called for side effects
}
\description{
Abort on column missingness
}
