% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{transition_model_version_stage}
\alias{transition_model_version_stage}
\title{Transition ModelVersion Stage}
\usage{
transition_model_version_stage(
  name,
  version,
  stage,
  archive_existing_versions = FALSE,
  client
)
}
\arguments{
\item{name}{Name of the registered model.}

\item{version}{Model version number.}

\item{stage}{Transition \code{model_version} to this tage.}

\item{archive_existing_versions}{(Optional)}

\item{client}{An MLFlow client. Will be auto-generated if omitted.}
}
\description{
Transition a model version to a different stage.
}
