% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{create_experiment}
\alias{create_experiment}
\title{Create Experiment}
\usage{
create_experiment(
  name,
  artifact_location = "",
  client = mlflow_client(),
  tags = list()
)
}
\arguments{
\item{name}{The name of the experiment to create.}

\item{artifact_location}{Location where all artifacts for this experiment are stored. If
not provided, the remote server will select an appropriate default.}

\item{client}{An MLFlow client. If not provided, the client is sourced from the \code{MLFLOW_TRACKING_URI} environment variable.}

\item{tags}{Experiment tags to set on the experiment upon experiment creation.}
}
\description{
Creates an MLflow experiment and returns its id.
}
