% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{create_model_version}
\alias{create_model_version}
\title{Create a model version}
\usage{
create_model_version(name, source, run_id, tags, run_link, description, client)
}
\arguments{
\item{name}{Register model under this name.}

\item{source}{URI indicating the location of the model artifacts.}

\item{run_id}{MLflow run ID for correlation, if \code{source} was generated
by an experiment run in MLflow Tracking.}

\item{tags}{Additional metadata.}

\item{run_link}{MLflow run link - This is the exact link of the run that
generated this model version.}

\item{description}{Description for model version.}

\item{client}{An MLFlow client. Will be auto-generated if omitted.}
}
\description{
Create a model version
}
