% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{create_registered_model}
\alias{create_registered_model}
\title{Create registered model}
\usage{
create_registered_model(name, tags, description, client)
}
\arguments{
\item{name}{The name of the model to create.}

\item{tags}{Additional metadata for the registered model (Optional).}

\item{description}{Description for the registered model (Optional).}
}
\description{
Creates a new registered model in the model registry
}
