% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{create_run}
\alias{create_run}
\title{Create an MLFlow run}
\usage{
create_run(start_time, tags, experiment_id, client)
}
\arguments{
\item{start_time}{The start time of the run. Defaults to the current time.}

\item{tags}{Additional tags to supply for the run}

\item{experiment_id}{The ID of the experiment to register the run under.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\description{
Create an MLFlow run
}
