% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{delete_model_version}
\alias{delete_model_version}
\title{Delete a model version}
\usage{
delete_model_version(name, version, client = mlflow_client())
}
\arguments{
\item{name}{Name of the registered model.}

\item{version}{Model version number.}

\item{client}{An MLFlow client. Will be auto-generated if omitted.}
}
\description{
Delete a model version
}
