% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{delete_registered_model}
\alias{delete_registered_model}
\title{Delete registered model}
\usage{
delete_registered_model(name, client = NULL)
}
\arguments{
\item{name}{The name of the model to delete}
}
\description{
Deletes an existing registered model by name
}
