% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{get_experiment}
\alias{get_experiment}
\title{Get Experiment}
\usage{
get_experiment(experiment_id, name, client = NULL)
}
\arguments{
\item{experiment_id}{ID of the experiment.}

\item{name}{The experiment name. Only one of \code{name} or \code{experiment_id} should be specified.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\description{
Gets metadata for an experiment and a list of runs for the experiment. Attempts to obtain the
active experiment if both \code{experiment_id} and \code{name} are unspecified.
}
