% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{get_experiment_id}
\alias{get_experiment_id}
\title{Get Experiment ID}
\usage{
get_experiment_id(experiment_name, client = mlflow_client())
}
\arguments{
\item{experiment_name}{The experiment name.}

\item{client}{An MLFlow client.}
}
\description{
Makes a call to \code{get_experiment} and returns just the ID.
}
