% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{get_latest_versions}
\alias{get_latest_versions}
\title{Get latest model versions}
\usage{
get_latest_versions(
  name,
  stages = list("None", "Archived", "Staging", "Production"),
  client = mlflow_client()
)
}
\arguments{
\item{name}{Name of the model.}

\item{stages}{A list of desired stages. If the input list is missing, return
latest versions for ALL_STAGES.}

\item{client}{An MLFlow client. Will be auto-generated if omitted.}
}
\description{
Retrieves a list of the latest model versions for a given model.
}
