% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{get_model_version}
\alias{get_model_version}
\title{Get a model version}
\usage{
get_model_version(name, version, client)
}
\arguments{
\item{name}{Name of the registered model.}

\item{version}{Model version number.}
}
\description{
Get a model version
}
