% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightMLFlow-package.R
\docType{package}
\name{lightMLFlow-package}
\alias{lightMLFlow}
\alias{lightMLFlow-package}
\title{lightMLFlow: A lightweight R wrapper for the MLFlow REST API}
\description{
This package provides a lightweight R wrapper for the MLFlow
    REST API. It makes some assumtions (notably, that you already have
    MLFlow installed wherever this package is being used), but, in return,
    you get some benefits: A simpler API, not needing `reticulate`,
    relying on `aws.s3` where possible, etc.
}
\author{
\strong{Maintainer}: Matt Kaye \email{mrkaye97@gmail.com}

}
\keyword{internal}
