% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{log_params}
\alias{log_params}
\title{Log Parameters}
\usage{
log_params(params, run_id, client)
}
\arguments{
\item{params}{A dataframe of params to log, containing the following columns: "key", "value".
This dataframe cannot contain any missing ('NA') entries.}

\item{run_id}{A run uuid. Automatically inferred if a run is currently active.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\description{
Logs parameters for a run. Examples are params and hyperparams
used for ML training, or constant dates and values used in an ETL pipeline.
A param is a STRING key-value pair. For a run, a single parameter is allowed
to be logged only once.
}
