% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{rename_experiment}
\alias{rename_experiment}
\title{Rename Experiment}
\usage{
rename_experiment(new_name, experiment_id = NULL, client = NULL)
}
\arguments{
\item{new_name}{The experiment's name will be changed to this. The new name must be unique.}

\item{experiment_id}{ID of the associated experiment. This field is required.}

\item{client}{An MLFlow client. Defaults to \code{NULL} and will be auto-generated.}
}
\description{
Renames an experiment.
}
