% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{rename_registered_model}
\alias{rename_registered_model}
\title{Rename a registered model}
\usage{
rename_registered_model(name, new_name, client = NULL)
}
\arguments{
\item{name}{The current name of the model.}

\item{new_name}{The new name for the model.}
}
\description{
Renames a model in the Model Registry.
}
