% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{set_experiment_tag}
\alias{set_experiment_tag}
\title{Set Experiment Tag}
\usage{
set_experiment_tag(key, value, experiment_id, client)
}
\arguments{
\item{key}{Name of the tag. All storage backends are guaranteed to support
key values up to 250 bytes in size. This field is required.}

\item{value}{String value of the tag being logged. All storage backends are
guaranteed to support key values up to 5000 bytes in size. This field is required.}

\item{experiment_id}{ID of the experiment.}

\item{client}{An MLFlow client. Will be auto-generated if not specified.}
}
\description{
Sets a tag on an experiment with the specified ID. Tags are experiment metadata that can be updated.
}
