% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{update_model_version}
\alias{update_model_version}
\title{Update model version}
\usage{
update_model_version(name, version, description, client)
}
\arguments{
\item{name}{Name of the registered model.}

\item{version}{Model version number.}

\item{description}{Description of this model version.}

\item{client}{An MLFlow client. Will be auto-generated if omitted.}
}
\description{
Updates a model version
}
