% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{update_registered_model}
\alias{update_registered_model}
\title{Update a registered model}
\usage{
update_registered_model(name, description = "", client = mlflow_client())
}
\arguments{
\item{name}{The name of the registered model.}

\item{description}{The updated description for this registered model.}

\item{client}{An MLFlow client. Will be auto-generated if omitted.}
}
\value{
Metadata on the registered model, including the \code{name} and the \code{creation_timestamp} and \code{last_updated_timestamp}
}
\description{
Updates a model in the Model Registry.
}
