% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{update_registered_model}
\alias{update_registered_model}
\title{Update a registered model}
\usage{
update_registered_model(name, description, client = NULL)
}
\arguments{
\item{name}{The name of the registered model.}

\item{description}{The updated description for this registered model.}
}
\description{
Updates a model in the Model Registry.
}
