% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with.mlflow_run}
\alias{with.mlflow_run}
\title{A \code{with} wrapper for MLFlow runs}
\usage{
\method{with}{mlflow_run}(data, expr, ...)
}
\value{
No return value. Called for side effects
}
\description{
Adds some error handling on exit
}
